requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.7
django-allauth==0.34.0
django-cors-headers==2.1.0
django-rest-auth==0.9.2
django-rest-swagger==2.1.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
psycopg2==2.7.3.2
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.12.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
django-model-utils==3.0.0
whitenoise==3.3.1
pycountry==17.9.23
django-redis-cache==1.7.1
django-storages==1.6.6
dropbox==8.7.1
django-filter==1.1.0
django-autoslug==1.9.3
django-redis-cache==1.7.1
rest-framework-cache==0.1
Pillow==4.3.0
django-mptt==0.9.0
